fix: dynamic room types in admin/student form, add DELETE /remove-room/{id} endpoint, seed default room types

+++ StableRoomie/src/main/java/in/edu/ssn/hostel/repo/roomRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface roomRepo extends JpaRepository<Rooms, Integer> {
+public interface roomRepo extends JpaRepository<Rooms, Long> {
 }
